Prepare 1.0.2 release. Removed other subprojects. Reformatted. Use new .github workflows. Added changelog. Export `charCell`.
