+++ .github/CODEOWNERS
@@ -1,0 +1,1 @@
+@byteverse/l3c
