+++ .github/workflows/build.yaml
@@ -1,0 +1,11 @@
+name: build
+on:
+  pull_request:
+    branches:
+      - "*"
+
+jobs:
+  call-workflow:
+    uses: byteverse/.github/.github/workflows/build-matrix.yaml@main
+    with:
+      cabal-file: lucid-colonnade.cabal
