+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+# Revision history for lucid-colonnade
+
+## 1.0.2 -- 2024-03-06
+
+* Update package metadata.
+* Export `charCell`.
