--- blaze-colonnade/blaze-colonnade.cabal
@@ -1,38 +1,0 @@
-name: blaze-colonnade
-version: 1.2.2.1
-synopsis: blaze-html backend for colonnade
-description:
-  This library provides a backend for using blaze-html with colonnade.
-  It generates standard HTML tables with `<table>`, `<tbody>`, `<thead>`,
-  `<tr>`, `<th>`, and `<td>`.
-homepage: https://github.com/andrewthad/colonnade#readme
-license: BSD3
-license-file: LICENSE
-author: Andrew Martin
-maintainer: andrew.thaddeus@gmail.com
-copyright: 2017 Andrew Martin
-category: web
-build-type: Simple
-cabal-version: >=1.10
-
--- Note: There is a dependency on profunctors whose only
--- purpose is to make doctest work correctly. Since this
--- library transitively depends on profunctors anyway,
--- this is not a big deal.
-
-library
-  hs-source-dirs:      src
-  exposed-modules:
-    Text.Blaze.Colonnade
-  build-depends:
-      base >= 4.8 && < 5
-    , colonnade >= 1.1 && < 1.3
-    , blaze-markup >= 0.7 && < 0.9
-    , blaze-html >= 0.8 && < 0.10
-    , profunctors >= 5.0 && < 5.7
-    , text >= 1.2 && < 2.1
-  default-language:    Haskell2010
-
-source-repository head
-  type:     git
-  location: https://github.com/andrewthad/colonnade
