--- build
@@ -1,16 +1,0 @@
-#!/bin/bash
-set -e
-
-# To use this script on Ubuntu, you will need to first run the following:
-# 
-# sudo apt install ghc-7.4.2 ghc-7.6.3 ghc-7.8.4 ghc-7.10.3 ghc-8.0.2 ghc-8.2.2 ghc-8.4.3 ghc-8.6.1
-
-declare -a ghcs=("7.10.3" "8.0.2" "8.2.2" "8.4.4" "8.6.5")
-
-## now loop through the above array
-for g in "${ghcs[@]}"
-do
-   cabal new-build --builddir "dist-$g" -w "/opt/ghc/bin/ghc-$g" colonnade
-   cabal new-build --builddir "dist-$g" -w "/opt/ghc/bin/ghc-$g" siphon
-done
-
