--- cabal.project
@@ -1,5 +1,0 @@
-packages: ./colonnade
-          ./blaze-colonnade
-          ./lucid-colonnade
-          ./siphon
-          ./yesod-colonnade
