--- colonnade/colonnade.cabal
@@ -1,50 +1,0 @@
-name: colonnade
-version: 1.2.0.2
-synopsis: Generic types and functions for columnar encoding and decoding
-description:
-  The `colonnade` package provides a way to talk about
-  columnar encodings and decodings of data. This package provides
-  very general types and does not provide a way for the end-user
-  to actually apply the columnar encodings they build to data.
-  Most users will also want to one a companion packages
-  that provides (1) a content type and (2) functions for feeding
-  data into a columnar encoding:
-  .
-  * <https://hackage.haskell.org/package/lucid-colonnade lucid-colonnade> for `lucid` html tables
-  .
-  * <https://hackage.haskell.org/package/blaze-colonnade blaze-colonnade> for `blaze` html tables
-  .
-  * <https://hackage.haskell.org/package/reflex-dom-colonnade reflex-dom-colonnade> for reactive `reflex-dom` tables
-  .
-  * <https://hackage.haskell.org/package/yesod-colonnade yesod-colonnade> for `yesod` widgets
-  .
-  * <http://hackage.haskell.org/package/siphon siphon> for encoding and decoding CSVs
-homepage: https://github.com/andrewthad/colonnade#readme
-license: BSD3
-license-file: LICENSE
-author: Andrew Martin
-maintainer: andrew.thaddeus@gmail.com
-copyright: 2016 Andrew Martin
-category: web
-build-type: Simple
-cabal-version: >=1.10
-
-library
-  hs-source-dirs:      src
-  exposed-modules:
-    Colonnade
-    Colonnade.Encode
-  build-depends:
-      base >= 4.12 && < 5
-    , contravariant >= 1.2 && < 1.6
-    , vector >= 0.10 && < 0.14
-    , text >= 1.0 && < 2.1
-    , bytestring >= 0.10 && < 0.12
-    , profunctors >= 5.0 && < 5.7
-    , semigroups >= 0.18.2 && < 0.21
-  default-language:    Haskell2010
-  ghc-options: -Wall
-
-source-repository head
-  type:     git
-  location: https://github.com/andrewthad/colonnade
