--- geolite-csv/data/small/GeoLite2-City-Blocks-IPv4.csv
@@ -1,11 +1,0 @@
-network,geoname_id,registered_country_geoname_id,represented_country_geoname_id,is_anonymous_proxy,is_satellite_provider,postal_code,latitude,longitude,accuracy_radius
-24.165.56.0/22,5848280,6252001,,0,0,96746,22.0837,-159.3553,10
-78.146.173.128/25,2655583,2635167,,0,0,DL14,54.6500,-1.6667,20
-121.211.108.0/23,2160386,2077456,,0,0,2040,-33.8833,151.1500,5
-69.74.43.16/30,6252001,6252001,,0,0,,37.7510,-97.8220,1000
-77.128.35.136/30,3034803,3017382,,0,0,57450,49.0667,6.8333,20
-90.54.234.0/24,2977062,3017382,,0,0,49320,47.3944,-0.4357,50
-77.193.41.175/32,3018587,3017382,,0,0,78810,48.8700,1.9740,1
-58.188.32.0/24,1861060,1861060,,0,0,,35.6900,139.6900,500
-87.81.232.0/24,2635167,2635167,,0,0,,51.4964,-0.1224,200
-88.191.56.0/22,2988507,3017382,,0,0,75001,48.8667,2.3333,500
