--- geolite-csv/geolite-csv.cabal
@@ -1,52 +1,0 @@
-name:                geolite-csv
-version:             0.2
-synopsis:            Geolite CSV Parser
-description:         Please see README.md
-homepage:            https://github.com/andrewthad/colonnade
-license:             BSD3
-license-file:        LICENSE
-author:              Andrew Martin
-maintainer:          andrew.thaddeus@gmail.com
-copyright:           2016 Andrew Martin
-category:            web
-build-type:          Simple
--- extra-source-files:
-cabal-version:       >=1.10
-
-library
-  hs-source-dirs:      src
-  exposed-modules:
-    Geolite.Types
-    Geolite.Csv
-  build-depends:
-      base >= 4.7 && < 5
-    , colonnade
-    , siphon
-    , ip   >= 0.8.4
-    , text
-    , pipes
-  default-language:    Haskell2010
-
-test-suite geolite-csv-test
-  type:                exitcode-stdio-1.0
-  hs-source-dirs:      test
-  main-is:             Spec.hs
-  build-depends:
-      base
-    , geolite-csv
-    , siphon
-    , colonnade
-    , test-framework
-    , text
-    , pipes
-    , HUnit
-    , test-framework-hunit
-    , pipes-bytestring
-    , pipes-text
-    , directory
-  ghc-options:         -threaded -rtsopts -with-rtsopts=-N
-  default-language:    Haskell2010
-
-source-repository head
-  type:     git
-  location: https://github.com/andrewthad/colonnade
