+++ lucid-colonnade.cabal
@@ -1,0 +1,36 @@
+cabal-version:   3.0
+name:            lucid-colonnade
+version:         1.0.2
+synopsis:        Helper functions for using lucid with colonnade
+description:     Helper functions for using lucid with colonnade.
+homepage:        https://github.com/byteverse/lucid-colonnade
+bug-reports:     https://github.com/byteverse/lucid-colonnade/issues
+license:         BSD-3-Clause
+license-file:    LICENSE
+author:          Andrew Martin
+maintainer:      amartin@layer3com.com
+copyright:       2017 Andrew Martin
+category:        web
+build-type:      Simple
+extra-doc-files: CHANGELOG.md
+tested-with:     GHC ==9.4.8 || ==9.6.3 || ==9.8.1
+
+common build-settings
+  default-language: Haskell2010
+  ghc-options:      -Wall -Wunused-packages
+  build-depends:    base >=4.8 && <5
+
+library
+  import:          build-settings
+  ghc-options:     -O2
+  hs-source-dirs:  src
+  exposed-modules: Lucid.Colonnade
+  build-depends:
+    , colonnade  >=1.1.1
+    , lucid      >=2.9
+    , text       >=1.2
+    , vector     >=0.10
+
+source-repository head
+  type:     git
+  location: git://github.com/byteverse/lucid-colonnade.git
