--- lucid-colonnade/lucid-colonnade.cabal
@@ -1,29 +1,0 @@
-name: lucid-colonnade
-version: 1.0.1
-synopsis: Helper functions for using lucid with colonnade
-description: Lucid and colonnade
-homepage: https://github.com/andrewthad/colonnade#readme
-license: BSD3
-license-file: LICENSE
-author: Andrew Martin
-maintainer: andrew.thaddeus@gmail.com
-copyright: 2017 Andrew Martin
-category: web
-build-type: Simple
-cabal-version: >=1.10
-
-library
-  hs-source-dirs: src
-  exposed-modules:
-    Lucid.Colonnade
-  build-depends:
-      base >= 4.8 && < 5
-    , colonnade >= 1.1.1 && < 1.3
-    , lucid >= 2.9 && < 3.0
-    , text >= 1.2 && < 2.1
-    , vector >= 0.10 && < 0.14
-  default-language: Haskell2010
-
-source-repository head
-  type:     git
-  location: https://github.com/andrewthad/colonnade
