--- projects/cabal-8.0.2.project
@@ -1,4 +1,0 @@
-packages: ./colonnade
-          ./blaze-colonnade
-          ./lucid-colonnade
-          ./yesod-colonnade
