--- projects/cabal-8.2.2.project
@@ -1,4 +1,0 @@
-packages: ./colonnade
-          ./blaze-colonnade
-          ./lucid-colonnade
-          ./yesod-colonnade
