--- projects/cabal-8.4.3.project
@@ -1,3 +1,0 @@
-packages: ./colonnade
-          ./blaze-colonnade
-          ./lucid-colonnade
