--- siphon/CHANGELOG.md
@@ -1,9 +1,0 @@
-# Revision history for siphon
-
-## 0.8.2.0 -- 2022-??-??
-
-* Add
-
-## 0.8.1.2 -- 2021-10-25
-
-* Correct handling of CRLF.
