--- siphon/siphon.cabal
@@ -1,58 +1,0 @@
-cabal-version: 3.0
-name: siphon
-version: 0.8.2.0
-synopsis: Encode and decode CSV files
-description: Please see README.md
-homepage: https://github.com/andrewthad/colonnade#readme
-license: BSD-3-Clause
-license-file: LICENSE
-author: Andrew Martin
-maintainer: andrew.thaddeus@gmail.com
-copyright: 2016 Andrew Martin
-category: web
-build-type: Simple
-extra-source-files: CHANGELOG.md
-
-library
-  hs-source-dirs: src
-  exposed-modules:
-    Siphon
-    Siphon.Types
-  build-depends:
-      base >= 4.8 && < 5
-    , colonnade >= 1.2 && < 1.3
-    , text >= 1.0 && < 2.1
-    , bytestring
-    , vector
-    , streaming >= 0.1.4 && < 0.3
-    , attoparsec
-    , transformers >= 0.4.2 && < 0.8
-    , semigroups >= 0.18.2 && < 0.21
-  default-language: Haskell2010
-
-test-suite test
-  type: exitcode-stdio-1.0
-  hs-source-dirs: test
-  main-is: Test.hs
-  build-depends:
-      base
-    , HUnit
-    , QuickCheck
-    , bytestring
-    , colonnade
-    , contravariant
-    , either
-    , pipes
-    , profunctors
-    , siphon
-    , streaming
-    , test-framework
-    , test-framework-hunit
-    , test-framework-quickcheck2
-    , text
-    , vector
-  default-language: Haskell2010
-
-source-repository head
-  type: git
-  location: https://github.com/andrewthad/colonnade
