--- siphon/src/Siphon/ByteString/Char8.hs
@@ -1,1 +1,0 @@
-module Siphon.ByteString.Char8 where
