--- siphon/src/Siphon/Content.hs
@@ -1,8 +1,0 @@
-module Siphon.Content
-  ( byteStringChar8
-  , text
-  ) where
-
-import Siphon.Internal (byteStringChar8)
-import Siphon.Internal.Text (text)
-
