--- yesod-colonnade/yesod-colonnade.cabal
@@ -1,33 +1,0 @@
-cabal-version: 2.0
-name: yesod-colonnade
-version: 1.3.0.2
-synopsis: Helper functions for using yesod with colonnade
-description: Yesod and colonnade
-homepage: https://github.com/andrewthad/colonnade#readme
-license: BSD3
-license-file: LICENSE
-author: Andrew Martin
-maintainer: andrew.thaddeus@gmail.com
-copyright: 2018 Andrew Martin
-category: web
-build-type: Simple
-
-library
-  hs-source-dirs: src
-  exposed-modules:
-    Yesod.Colonnade
-  build-depends:
-      base >= 4.9.1 && < 4.18
-    , colonnade >= 1.2 && < 1.3
-    , yesod-core >= 1.6 && < 1.7
-    , conduit >= 1.3 && < 1.4
-    , conduit-extra >= 1.3 && < 1.4
-    , text >= 1.0 && < 2.1
-    , blaze-markup >= 0.7 && < 0.9
-    , blaze-html >= 0.8 && < 0.10
-    , yesod-elements >= 1.1 && < 1.2
-  default-language: Haskell2010
-
-source-repository head
-  type: git
-  location: https://github.com/andrewthad/colonnade
